pyproject.toml
[project]
name = "plumkdocs"
version = "0.0.5"
description = "Hacky mkdocs documentation for plum-dispatch"
authors = [
    {name = "Antonio Stanziola", email = "stanziola.antonio@gmail.com"}
]
license = {text = "LGPL-3.0-only"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "plum-dispatch>=2.6.0,<3",
    "griffe>=1.15.0",
    "markdown>=3.10.0",
    "mkdocs-material>=9.7.0",
    "mkdocstrings>=1.0.0",
    "mkdocstrings-python>=2.0.0,<3",
    "mkdocs-macros-plugin>=1.5.0",
    "pygments>=2.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.0",
    "coverage[toml]>=7.13.0",
    "ruff>=0.14.0",
    "beartype>=0.22.0",
    "python-semantic-release>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/astanziola/plumkdocs"
"Bug Tracker" = "https://github.com/astanziola/plumkdocs/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["plumkdocs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=plumkdocs",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["plumkdocs"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/fixtures/*.py" = ["F811"]  # Allow redefinition for plum dispatch
"example_module/**/*.py" = ["F811", "W293", "W291"]  # Allow redefinition and whitespace in examples

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
build_command = "python -m build"
major_on_zero = false
tag_format = "v{version}"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "^chore",
    "^ci",
    "^docs(?!:)",
    "^style",
    "^test",
    "^refactor(?!:)",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease = false
